+++ CHANGELOG.md
@@ -33,6 +33,7 @@
 ### Added
 
 - Alternative titles of a work : search songs and works upon the alternative title of a work.
+- Reorder of the playlist
 
 ## 1.2.0 - 2018-06-03